requirements.txt
## The following requirements were added by pip freeze:
c3pyo==0.0.4.1
click==6.7
dominate==2.3.1
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
pyserial==3.3
visitor==0.1.3
Werkzeug==0.12.2
